backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-CORS>=4.0
PyMySQL>=1.0.0
python-dotenv
gunicorn>=21.0.0